requirements.txt
PyOpenGL
WxPython
numpy
pyrr
numpy-stl
pyquaternion
pillow
triangle